+++ api/v1/server/handlers/users/reject_invite.go
@@ -8,13 +8,13 @@ import (
 
 	"github.com/hatchet-dev/hatchet/api/v1/server/oas/apierrors"
 	"github.com/hatchet-dev/hatchet/api/v1/server/oas/gen"
-	"github.com/hatchet-dev/hatchet/pkg/repository"
-	"github.com/hatchet-dev/hatchet/pkg/repository/postgres/dbsqlc"
-	"github.com/hatchet-dev/hatchet/pkg/repository/postgres/sqlchelpers"
+	v1 "github.com/hatchet-dev/hatchet/pkg/repository"
+	"github.com/hatchet-dev/hatchet/pkg/repository/sqlchelpers"
+	"github.com/hatchet-dev/hatchet/pkg/repository/sqlcv1"
 )
 
 func (u *UserService) TenantInviteReject(ctx echo.Context, request gen.TenantInviteRejectRequestObject) (gen.TenantInviteRejectResponseObject, error) {
-	user := ctx.Get("user").(*dbsqlc.User)
+	user := ctx.Get("user").(*sqlcv1.User)
 	userId := sqlchelpers.UUIDToStr(user.ID)
 
 	// validate the request
@@ -31,7 +31,7 @@ func (u *UserService) TenantInviteReject(ctx echo.Context, request gen.TenantInv
 	}
 
 	// get the invite
-	invite, err := u.config.APIRepository.TenantInvite().GetTenantInvite(ctx.Request().Context(), inviteId)
+	invite, err := u.config.V1.TenantInvite().GetTenantInvite(ctx.Request().Context(), inviteId)
 
 	if err != nil {
 		return nil, err
@@ -48,17 +48,17 @@ func (u *UserService) TenantInviteReject(ctx echo.Context, request gen.TenantInv
 	}
 
 	// ensure invite is in a pending state
-	if invite.Status != dbsqlc.InviteLinkStatusPENDING {
+	if invite.Status != sqlcv1.InviteLinkStatusPENDING {
 		return gen.TenantInviteReject400JSONResponse(apierrors.NewAPIErrors("invite has already been used")), nil
 	}
 
 	// construct the database query
-	updateOpts := &repository.UpdateTenantInviteOpts{
-		Status: repository.StringPtr(string(dbsqlc.InviteLinkStatusREJECTED)),
+	updateOpts := &v1.UpdateTenantInviteOpts{
+		Status: v1.StringPtr(string(sqlcv1.InviteLinkStatusREJECTED)),
 	}
 
 	// update the invite
-	invite, err = u.config.APIRepository.TenantInvite().UpdateTenantInvite(ctx.Request().Context(), sqlchelpers.UUIDToStr(invite.ID), updateOpts)
+	invite, err = u.config.V1.TenantInvite().UpdateTenantInvite(ctx.Request().Context(), sqlchelpers.UUIDToStr(invite.ID), updateOpts)
 
 	if err != nil {
 		return nil, err