+++ frontend/app/src/pages/main/v1/workflow-runs-v1/components/v1/task-runs-columns.tsx
@@ -4,7 +4,7 @@ import {
 } from '../../../events/components/additional-metadata';
 import { V1RunStatus } from '../../../workflow-runs/components/run-statuses';
 import { DataTableColumnHeader } from '@/components/v1/molecules/data-table/data-table-column-header';
-import { DataTableRowActions } from '@/components/v1/molecules/data-table/data-table-row-actions';
+import { TableRowActions } from '@/components/v1/molecules/data-table/data-table-row-actions';
 import RelativeDate from '@/components/v1/molecules/relative-date';
 import { Duration } from '@/components/v1/shared/duration';
 import { Button } from '@/components/v1/ui/button';
@@ -320,8 +320,8 @@ export const columns: (
     id: 'actions',
     cell: ({ row }) => {
       return (
-        <DataTableRowActions
-          row={row}
+        <TableRowActions
+          row={row.original}
           actions={[
             {
               label: 'Copy Run Id',