+++ sdks/python/hatchet_sdk/clients/rest/models/tenant.py
@@ -24,7 +24,6 @@ from typing_extensions import Self
 
 from hatchet_sdk.clients.rest.models.api_resource_meta import APIResourceMeta
 from hatchet_sdk.clients.rest.models.tenant_environment import TenantEnvironment
-from hatchet_sdk.clients.rest.models.tenant_ui_version import TenantUIVersion
 from hatchet_sdk.clients.rest.models.tenant_version import TenantVersion
 
 
@@ -47,9 +46,6 @@ class Tenant(BaseModel):
         alias="alertMemberEmails",
     )
     version: TenantVersion = Field(description="The version of the tenant.")
-    ui_version: Optional[TenantUIVersion] = Field(
-        default=None, description="The UI of the tenant.", alias="uiVersion"
-    )
     environment: Optional[TenantEnvironment] = Field(
         default=None, description="The environment type of the tenant."
     )
@@ -60,7 +56,6 @@ class Tenant(BaseModel):
         "analyticsOptOut",
         "alertMemberEmails",
         "version",
-        "uiVersion",
         "environment",
     ]
 
@@ -127,7 +122,6 @@ class Tenant(BaseModel):
                 "analyticsOptOut": obj.get("analyticsOptOut"),
                 "alertMemberEmails": obj.get("alertMemberEmails"),
                 "version": obj.get("version"),
-                "uiVersion": obj.get("uiVersion"),
                 "environment": obj.get("environment"),
             }
         )