+++ sdks/python/hatchet_sdk/worker/runner/runner.py
@@ -100,16 +100,17 @@ class Runner:
         self.dispatcher_client = DispatcherClient(self.config)
         self.workflow_run_event_listener = RunEventListenerClient(self.config)
         self.workflow_listener = PooledWorkflowRunListener(self.config)
-        self.runs_client = RunsClient(
-            config=self.config,
-            workflow_run_event_listener=self.workflow_run_event_listener,
-            workflow_run_listener=self.workflow_listener,
-        )
         self.admin_client = AdminClient(
             self.config,
             self.workflow_listener,
             self.workflow_run_event_listener,
-            self.runs_client,
+        )
+
+        self.runs_client = RunsClient(
+            config=self.config,
+            workflow_run_event_listener=self.workflow_run_event_listener,
+            workflow_run_listener=self.workflow_listener,
+            admin_client=self.admin_client,
         )
         self.event_client = EventClient(self.config)
         self.durable_event_listener = DurableEventListener(self.config)
@@ -149,7 +150,9 @@ class Runner:
             self.running_tasks.add(t)
             t.add_done_callback(lambda task: self.running_tasks.discard(task))
 
-    def step_run_callback(self, action: Action) -> Callable[[asyncio.Task[Any]], None]:
+    def step_run_callback(
+        self, action: Action, t: Task[TWorkflowInput, R]
+    ) -> Callable[[asyncio.Task[Any]], None]:
         def inner_callback(task: asyncio.Task[Any]) -> None:
             self.cleanup_run_id(action.key)
 
@@ -173,7 +176,11 @@ class Runner:
                     )
                 )
 
-                log_with_level = logger.info if should_not_retry else logger.exception
+                # log as info if we're going to retry or we explicitly should _not_ retry
+                # so that e.g. Sentry does not get reported multiple exceptions from multiple retries of a single task
+                log_as_info = should_not_retry or action.retry_count < t.retries
+
+                log_with_level = logger.info if log_as_info else logger.exception
 
                 log_with_level(
                     f"failed step run: {action.action_id}/{action.step_run_id}\n{exc.serialize(include_metadata=False)}"
@@ -397,7 +404,7 @@ class Runner:
                 self.async_wrapped_action_func(context, action_func, action)
             )
 
-            task.add_done_callback(self.step_run_callback(action))
+            task.add_done_callback(self.step_run_callback(action, action_func))
             self.tasks[action.key] = task
 
             task_count.increment()